requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dateutil==2.8.2
newsapi-python==0.2.6
colorama==0.4.6
google-generativeai>=0.3.0